Move metrics config structs from collectd to confgenerator/config and make them pointers. Give some structs more precise names.
