+++ confgenerator/testdata/invalid/linux/metrics-processor_unsupported_type/golden_error
@@ -1,2 +1,2 @@
 the agent config file is not valid YAML. detailed error: yaml: unmarshal errors:
-  line 6: field processors not found in type collectd.Metrics
+  line 6: field processors not found in type config.Metrics
