+++ fluentbit/conf/conf.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-//Package conf provides data structures to represent and generate fluentBit configuration.
+// Package conf provides data structures to represent and generate fluentBit configuration.
 package conf
 
 import (
